requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
attrs==22.1.0
    # via pytest
certifi==2022.9.24
    # via requests
charset-normalizer==2.1.1
    # via requests
exceptiongroup==1.0.4
    # via pytest
idna==3.4
    # via requests
iniconfig==1.1.1
    # via pytest
packaging==21.3
    # via pytest
pluggy==1.0.0
    # via pytest
pyparsing==3.0.9
    # via packaging
pytest==7.2.0
    # via -r requirements.in
python-dotenv==0.21.0
    # via -r requirements.in
requests==2.28.1
    # via -r requirements.in
tomli==2.0.1
    # via pytest
urllib3==1.26.13
    # via requests
